Added new Migration

+++ Smart-Garage/Models/ViewModel/VisitViewModel.cs
@@ -8,9 +8,9 @@ namespace Smart_Garage.Models.ViewModel
         public int Id { get; set; }
 
         public DateTime Date { get; set; }
-        public int TotalPrice { get; set; } // in BGN
-        public int ServicesTotalPrice { get; set; } // in BGN
-        public int PartsTotalPrice { get; set; } // in BGN
+        public double TotalPrice { get; set; } // in BGN
+        public double ServicesTotalPrice { get; set; } // in BGN
+        public double PartsTotalPrice { get; set; } // in BGN
         public VehicleViewModel Vehicle { get; set; }
         public IList<ServiceInstanceViewModel> ServiceInstances { get; set; }
     }
